Committing two main things here. Both seem to be "working" in that the model runs and they do basically what they are supposed to be doing, but testing is definitely still ongoing!:
(1) putting the climate-interactive CH4 wetland emissions back in the model
[switched on with #define INTERACTIVE_WETLANDS_CH4]. Alterations to:
TRACERS_SPECIAL_Shindell.f
TRACERS_DRV.f
Please see the E001TdsM23 sample rundeck for the necessary input
files and parameters. For the moment, we recommend using nn_or_zon=1
One important note is to set ns_wet to the number of the CH4
surface source to be altered by the wetlands scheme!

(2) putting N.Unger's old implementation of climate-interactive
isoprene emissions back in the model.
[switched on with #define BIOGENIC_EMISSIONS]
Alterations to:
ALLOC_DRV.f
BIOGENIC_EMISSIONS.f [newly checked-in]
GHY_DRV.f
PBL.f
SURFACE.f
TRACERS_DRV.f
TRDIAG_COM.f
We realize that Nadine is not happy with this implementation and
has a totally new version coming, but some people needed to use
this in the mean time. So I am committed to avoid having to manually
merge it back in later. But this might all be replaced by her.