Moved reading of command line options and rundeck parameters to the very beginning of the main program. This should allow the use of rundeck parameters (those specified in "&&PARAMETERS" block) before the call to "input", for example to specify J_TARG or some dimensions before allocation routines are called.
One should remember though that before the call to "input" the
database contains only parameters which were specified in the
current rundeck (i.e. in "I"-file). The parameters which were
saved in the restart file are not available yet. In most cases
that shouldn't be a problem. There is though one case where
one should be careful: pdE script creates its own "I" file
with empty &&PARAMETERS block. That means that for post-processing
rundeck parameters are not available untill the call to "input".
If that creates a problem we can modify pdE so that it copies
&&PARAMETERS from current I file of from saved rundeck...

Of course a more robust way would be to modify input algorithm
so that first two records of restart file are read before the
call to init_app ...
But that again may be messy because during the post-processing
we read multiple files...