This is a preliminary version of parallelized obio model. The model was tested only for the flags
#define TRACERS_ON                  ! include tracers code
#define TRACERS_GASEXCH_Natassa     ! special tracers to be passed to ocean
#define TRACERS_GASEXCH_CO2_Natassa ! special tracers to be passed to ocean
#define TRACERS_OceanBiology        ! Watson Gregg's ocean bio-geo-chem model

and it produces the same results on 1 and 4 cpu's. Restart was not tested
yet and may require some work. Also, OBIO_RAD_coupling was not tested at the
moment.

The model still requires a lot of cleaning/optimizing. Global dimensions
of arrays should be replaced with local ones, some esmf_bcast() calls
should be replaced with scatter, unnecessary gather/scatter should be
removed...

Natassa, this includes the latest changes you have given to me.
For a reference I'm going to tag the version that is working for me
with "obio_mpi_1" (my non-ocean part is ~ 1 week old, though I doubt
that updating to the latest repository code will break anything).