Introduction of new versions of routines COSZT/COSZS which can handle a non-latlon grid.
To use, #define NEW_COSZ in the rundeck.  This causes RAD_DRV.f
to #include cosz_new.inc - the #include mechanism can be
eliminated later.

For now, SINJ and COSJ are still used until the geometric information
for alternative grids is introduced.

Verification of correctness w.r.t. "cut-off values of cosz" remains
to be done.