I had noticed using CLOUDS2.f, I was getting negative tracers. Dorothy Koch believes she found the problem, the fix is what I am committing here. She says:
"Someone tried to combine some codes and replaced a get_wash
with a get_cond when determining TMFAC2. Then this statement is
not correct:

 TMFAC2=THWASH/TM(L,N)

since THWASH is not [defined] by GET_COND_FACTOR only by GET_WASH."

In case others want to test it, I have left the code changes marked
with the cdmk2 comments.