Replaced model/DOMAIN_DECOMP.f with an empty shell which includes the corresponding file from ESMF_Interface. This is to avoid code duplication and related confusion. DOMAIN_DECOMP.f needs to be in ESMF_Interface for the structure with subdirs to work. If you need to edit this file please edit ESMF_Interface/DOMAIN_DECOMP.f.
I did a hack and moved this file by hand inside the repository
to preserve its modification history. This shouldn't break anything,
but if you feel that something is wrong with this file after update
try to delete it and do "cvs update" again.

If you plan to modify it, I suggest you do "cvs update" first.
Otherwise your changes may be lost due to conflict in version
numbers.