Apparently there was an error in FBBphotosynthesis.f . "Jc" used Ko while Ci_Jc used pspar%Ko . The error may have been not that big but it was creating incompatible assimilation fluxes which could crash the model.
I got rid of calls to Jc and Je since they were duplicating computations
already done in ci_cubic. Can we eliminate these functions completely?
They are used in something called Farquhar_standalone . Do we need
this? If there are functions here which are needed only for some
extensive debugging, can they be moved to a separate file?
It seems like we have a lot of "dead" code here.

Also, ci seems to be converted from umol to Pa and then back to umol.
Can we keep it in umol everywhere?

Eliminated old version of "ci_cubic".

Added some checks for LAI==0 to avoid zero-divisions.