Additional connections between GSFC's FVcube and modelE:
* Instanciated Max' 2d domain decomposition structure (dd2d) and glued it to the dist_grid derived type, passing subdomain bounds is,js,ie,je,... to dd2d.

* init_csgrid_debug now deprecated

* added dd2d_utils.f, pario_nc.f and pario_fbsa.f in main directory. These are links to the same files in Max' dd2d directory.

* added CUBE_GRID cpp flag in rules.mk.
cpp flags CUBE_GRID, CUBE_SPHERE and FVCUBED will be unified later.

* for testing the links between cubed sphere core, .modelErc has to contain (mostly copied from Rahman Syed's post):
   FVCORE_ROOT=/discover/nobackup/projects/map/GEOSGCM
   FVCUBED_ROOT=/discover/nobackup/projects/map/GEOSagcm-cubed
   MPPDIR=/discover/nobackup/projects/map/MPP/mpp_intel-9.1.052_scali-5
   BASELIBDIR=/usr/local/other/baselibs/ESMF222rp3_NetCDF362b6_9.1.052/Linux
   CUBE_GRID=YES

* add 2 lines in the E1fvcM20.R rundeck
  dd2d_utils pario_nc pario_fbsa
  regrid regrid_com regridinput

* gmake gcm RUN=E1fvcM20 ESMF=YES FVCUBED=YES
  gmake setup_nocomp RUN=E1fvcM20 NPES=6

Later on, we will need to think about a design that :
1) avoids duplication of information between dist_grid, domain2d and dd2d_grid,
2) DOMAIN_DECOMP (1d decomp.) and DD2D (2d decomp.) modules are separated.

To achieve this, it would be good to
a) maintain a unique logical grid (e.g. dist_grid), which would contain all the information of the existing dist_grid plus the information common to domain2d and dd2d_grid, and makes it available to both lat-lon (DOMAIN_DECOMP) and 2d (dd2d) modules. References to dd2d_grid in Max' dd2d module would then be replaced by references to dist_grid
b) move the instantiation of the dist_grid object from DOMAIN_DECOMP to a separate file