tracer code adjustments to i) use axyp(i,j) instead of dxyp(j), ii) use two dimensional domain decomposition (i=1,im => i=i_0,i_1), and iii) have tajln/tajls use function calls to be integrated.
This coding is preliminary and it should not be supposed that the tracers
will work perfectly on the CS as yet. Specifically, there are a number
of zonal mean calculations which require some adjustment.

Some of this code was tested, but some might not be in the configurations
I used. Please let me know ASAP if there are any compile or runtime issues with
this commit.