Fixes and tidies to CLOUDS2 code:
- update of energy conservation checks in CLOUDS2_DRV.f
- fix of energy losses in CLOUDS2 due to:
	* incorrect spec of HEAT1 (now applied to plume)
	* missed loops for application of HEAT1
- improved clarity for phase of moist convective condensate and
precipitation. This is defined based on the first plume for condensate,
and on the environment for precip. This makes a noticeable change in
the amount of moist convection and may need to be looked at.

- addition of some generalised functions for calculating condensation to
avoid duplication and occasional incorrect updating of T/Q (get_dq_evap,
get_dq_cond)

- tidying up of dead code - please do not leave in code that has been
superseded - it makes debugging/maintenance much harder

- improved comments - please add more informative comments to any coding
Note also, that the cvs system keeps good track of how added what when.
There is no need to additionally put such comments in the code.

CLOUDS2 now conserves energy to machine accuracy. Use the differences between
CHANGE OF TPE BY CONDSE and CHANGE OF ENRG WAT BY CONDSE to keep track of
this in future.