In subroutine DIAGIJt:
converted automatic global arrays aij1 and aij2 to allocatable
to allow the code to compile at 360x180xL40 horizontal resolution
with the latest setting for ktaijs.

Not 100% sure why this was necessary, but it seems the
authors of ifort (perhaps justifiably) did not envision
compile-time requests for 2.1 GB worth of automatic arrays
in one subroutine.

In any case, DIAGIJt should probably be redesigned to process
diagnostics one by one, as opposed to the current practice of
having the root PE temporarily store all the global-sized
instances of mostly independent output quantities.