Changes for coexistence of a cubed-sphere atmosphere and latlon ocean+seaice grids:
The DOMAIN_DECOMP_ATM module is only defined in this
file if the atmosphere is lat-lon; otherwise this
module is defined in a separate file.
The initialization of the atmospheric grid has been moved
from the init_app() routine to main(), so that init_grid()
can be obtained from the other file if necessary.

MPP+FVcubed preprocessor directives related to a
cubed-sphere atmosphere have been left in place for now,
but are set to undefined at the beginning of the file.

In the cubed sphere case, the definition of the
dist_grid derived type is taken from the dd2d_utils
module.  This facilitates plug compatibility
of DOMAIN_DECOMP_ATM and allows quantities on
both cubed and latlon grids to be read/written
from/to restart files using the same routines.