Yeonjoo and Igor:  I commented out the time variable from lsm_standalone.f, because the code does not compile with it.
Also, I changed a line in GHY.f from
real*8, save, time=0.d0
to
real*8 time

The code did not compile without these changes.


Igor:  I moved the development notes from the beginning of GHY.f to the end (I make note of this at the start of the file).
I can move it back if it is important.  I just moved it because the notes are a bit annoying to see everytime I open the file.
:-)