For grid portability and offline postprocessing, a number of aspects of IJ diagnostics have been changed.  The dust hasn't quite settled yet; let me know of any problems.
Special derived quantities are computed before
printing; almost all the code for this was moved
to DIAG.f.  A few grid-specific computations
remain in ij_mapk for quantities derived from
B-grid averages of A-grid fields.
Hopefully this averaging will be eliminated.

Time weighting for ratios and area weighting for
computation of spatial means are now treated the
same way.  Every accumulation which needs time
and/or area weighting must specify denom_ij = IJ_xxx,
where IJ_xxx is the index of the AIJ element
to use as the weight.  Some extra elements have
been added to AIJ for this purpose, and multiplication
by area fraction has been added to some accumulations.

Results differ for diagnostics over the soil
fraction in the variable-lakes case, as there was
no weighting by time-variable type fraction before.