Changes to budget-page diagnostics to facilitate offline postprocessing.  This is work in progress; let me know of any problems.
Computation of special derived quantities and surface
type composites was moved to a routine DIAGJ_PREP which
assembles a print-ready array AJ_OUT.  Since all AJ
accumulations are multiplied by surface type, all AJ_OUTs
save for surface type can be interpreted as ratios (like
albedo). This eliminates the special treatment of albedos
and mc cloud depth.  Quantities that are ratios are
tagged with IDEN_J = J_xxx where J_xxx is the index
of the relevant denominator.

Undefined albedos that were reported as 100% are now
reported as 0%.

Analogous changes were applied to the regions diagnostics.