Committing a simple attempt at biomass burning tracer emissions that react to the flammability of surface vegetation. **This is still in testing**, and it is turned on with #define ALTER_BIOMASS_BY_FIRE. #define CALCULATE_FLAMMABILITY must also be on.
As of now, it can be used in two cases:
(1) if you have any surface source that is tagged with the sector:
    'BBURN' in the rundeck, that will be altered with climate.
(2) if you are using the GFED 3D biomass burning files.
    (currently, the "base flammability" is from a run with
    prescribed SSTs during the 1997-2002 GFED time period).

Currently, this can only be used in the 2x2.5 model, because it
requires rundeck file:
BASE_FLAM=gsin/flammability_MON1997-2002_GFED.ijEflamBF40
But that's just until I made other resolution's input files.

It might seem odd to put this in TRACERS.f and TRACERS_DRV.f
(general tracer code), but the intention is that it can be used
by others e.g. without the Shindell tracers on. Unfortunately, I have
introduced another zonal mean that I will have to alter for the
cubic sphere and 2D decomposition cases. I will do those all at once.