More updates for grid independence and streamlined postprocessing.
Completed the transformation AIL -> AIJL: changed the names
of arrays and indices, and introduced metadata like ia_ijl,
scale_ijl et al.

AIJL is intended to contain only diagnostics whose accumulation
details are independent of the atmospheric dynamics grid staggering.
These can be either on model layers (IJL_xxx indices) or
constant-pressure levels (IJK_xxx indices).  All the diagnostics
that were written using POUT_IJL have been moved from the AIJK
array to AIJL.  A scaling routine IJLMAP was created which
loops over the contents of AIJL.
The constant-pressure specific humidity diagnostic was transferred
from AIJK to AIJL, and the relative humidity field in AIJK was
deleted.  The standing eddy calculation in DIAGJK performs
the interpolation of A-grid specific humidity to the B grid.
The model-layer temperature and relative humidity diagnostics
in the former AIL are now on constant-pressure levels in AIJL.
Scaling and titles in DIAGIL were altered to reflect this change.

The units of the AIJL accumulations for radiative and
moist convective heating are now per square meter; the
units in DIAGIL are now W/(m2*mb) rather than WATTS/DSIG.
Arrangements can be made to accomodate fans of the
latter unit.

The column temperature IJ diagnostic is now on the A grid.
MSU temperatures are now on the A grid and are computed
in the parallel section of the code.
An error in the surface temperature input to the B-grid MSU
calculation was corrected in the A-grid version.

The optional 3D moist convective heating rate diagnostics are
automatically scaled by air mass in IJLMAP to produce the
indicated units of W/(m2*mb).

To indicate that they are accumulated on the B grid,
the indices IJK_[U,V,T,DP] for AIJK were renamed to
IJK_[UB,VB,TB,DPB].