Removed geometric routines that now live elsewhere.
Changed the interface of interpolation routines to match
the needs of the routines in GCDIAGcs.f (only interpolate
to 1 latitude at a time).

Added a routine corner_fill which facilitates bilinear
interpolation near the corners of the cube.

Added a routine sumxpe_zonal which handles the mpi_allreduce
calls for zonal sums.