--> Adding program lightning. --> Including tuning dbparams: tune_lt_land and tune_lt_sea. --> listing (but commenting  #define CALCULATE_LIGHTNING)
Taking the Colin Price lightning parameterization out of the tracer
code and into its own program lightning.f. This is so it can be used
when tracers are off. You can turn it on with #define CALCULATE_LIGHTNING
or the usual #define TRACERS_SPECIAL_Shindell. Likewise moving it's
ij diags (flash rate and cloud-to-ground flash rate) out of the tracer
diags and into the gcm diags.

Some tuning may still be done (certainly for the 2x2.5 model resolution),
and these tuning parameters are made db-parameters so that they can be
altered in rundecks, based on resolution.