NEW_IO: Writing an "acc-file" after the first timestep during setup so that offline diagnostic routines can report on initial conditions.
Cubed sphere:
Added some comments to cdl_aij/cdl_aijl that facilitate
offline remapping from the cubed sphere to a latlon grid.
JM_BUDG is now a function of resolution, and the nominal
latitudes of the budget grid are filled in.

FYI:
Reducing JM_BUDG from 46 to 24 for C32 resolution causes a
COMPILE_WITH_TRAPS floating-point exception crash in the
FVcubed dycore for NPES=6 (NPES=24 runs OK).  Without the
COMPILE_WITH_TRAPS, prognostic arrays are identical on 6
and 24 PEs, and the "nonreproducible" diagnostic arrays
are acceptably close. Similar aborts in the FV dynamics
have been triggered by unrelated changes to the modelE
code.
Some hypotheses are:
(1) There is an uninitialized allocatable array somewhere,
not necessarily in the modelE code
(2) Something on the modelE side is reaching over and
corrupting the FV memory space, but in a way which does
not trip any array bounds checks.