Changes for cubed-sphere compatibility of LAKES.f.
Changed some loop bounds to always include halo rows in the cubed
sphere case.  Added have_south_pole/have_north_pole checks in
the get_dir function.

Discovered that the RATE array needs to be set even for cells whose
flow is leaving the local tile, and that this calculation needs
iflow,jflow to be set in such cases.  So, the lonlat_to_ij
function has been upgraded to return meaningful values for
points on adjacent tiles.
IJ pairs for points on adjacent tiles are reported in terms
of the local IJ space.  These changes allow the older version of
horzdist_2pts to work as well.