reversing earlier changes.
lam is integer and read in from a file
lam8 is real8 and used for reflectance calculations

please, define OBIO_RAD_COUPLING when
direct and diffuse radiation comes from the atmos model
and is NOT read in from a file.

In order to run case with radiation coupling to chlorophyll must
define OBIO_RAD_coupling *AND* either chl_from_seawifs for
prescribed chlorophyll or chl_from_obio for prognostic chlorophyll.