Changed rougness length for vegetation to higher values which are more close to the values in E31*_ia3 experiment. These new values are pretty much based on a rule (used by oter modeling groups)  z0 ~ h/20 for tall vegetation (trees)  z0 ~ h/10 for short vegetation (grass, crops, bushes) This is also consistent with values suggested in Monin & Yaglom for short vegetation: z0 ~ h/7.5 .
Original values from 1983 paper are left in the code but are commented
out. If one wants to run the model with those values one should
uncomment corresponding definition for z0_cover(:) in GHY_DRV.f .

Currently this algorithm is turned on only for

 ( defined ROUGHL_HACK ) && ( defined USE_ENT )

otherwise the roughness length is read from the file.