Changes so that conservation diagnostics work for the 1x1.25 ocean + 2x2.5 atm configuration, which has different atm/ocean domain decompositions unless using special numbers of PEs.  Forcing both the atmosphere and ocean conservation diagnostics to occupy the same CONSRV array means that CONSRV cannot be distributed unless atm grid == ocean grid.
Switched CONSRV_loc from being distributed to
global.  SUMXPE used in place of PACK_DATAJ.
Hacks to account for the fact that some elements
of CONSRV are instantaneous values (but are
nevertheless summed during pdE).

For the 1x1.25 ocean + 2x2.5 atm configuration, the
ocean parts of the CONSRV array will have roundoff
differences on different numbers of PEs, since
the latitude bands for CONSRV are split by the
ocean domain decomposition.

I have not tested with pdE.