For Ox tracer: added JL Ox chem production in troposphere and Ox chem destruction in troposphere diagnostics to compliment the full-atmosphere versions.
Removed a factor of byim in the scaling of JL CO and Ox production
and destruction by chemistry. Changed power of 10 factor too.

New method of biomass burning emissions based on
climate-sensetive flammability (Replacing
#ifdef ALTER_BIOMASS_BY_FIRE methodology with
#if def DYNAMIC_BIOMASS_BURNING methodology.)

** Right now, this seems to be producing sources that are far too big. **

But I tried to code this so that anyone can turn on a flammability-
sensitive biomass burning source(s) for their tracer, once they make
a correlation file between emissions and fire count, and listing it in the
rundeck as trname(n)_EPFC. We can provide instructions for this,
but might as well wait until the problem with the magnitude is solved
for our tracers.