Comments only.
Wanted to put into the log Drew's clarification of yesterday's
commit notes:

The main chemistry improvements were to fully conserve nitrogen
and to remove sensitivity to the location of the tropopause.

The "Altering the few remaining tunings of ss for N2O and CFC"
in TRCHEM_master.f, is "for photolysis window region"

And the "Changes to the total Br and Cl calculations" were to
fully conserve nitrogen when ClONO2 and BrONO2 are normalized
to CFC+background (which is the time-saving way of creating
the appropriate Cl and Br sources and sinks).