conservation under restart for the special case of interactive gas exchange.
Several arrays in obio_model are computed at the start of a day. Therefore if the model
is restarted from a time step other than the first timestep of the day, the results
differ after restart. A solution is to avoid restarting the model from hour.neq.0.
Eg. force setup for an entire day, not just an hour.
Also, use nssw=48 in the rundeck and a check inside obio_model to redo the hour=0 calculations.
Alternative would be to save the day_start arrays in a restart file, which would though
increase the size of fort.[12] too much.

If anyone has better suggestions, please let me know.