Particularly of interest to anyone running tracer dry deposition code that used the old Harvard CTM leaf area index routines. To summarize: This is a bug fix. Results will change. After running like this, please check your tracer dry dep part of budget.
This commit fixes two bugs:

(1) the first was a roundoff error upon restarts because interpolation of XYLAI was done slightly differently the first time step. This caused an irreproducibility. Thanks Max K.

(2) the second is a more serious long-standing reproducibility bug:  It seems to me that upon restarts and initial start, the wrong months were read for LAI files. Example: upon (re)starts in the first half of Jan, it would read JAN and FEB and interpolate, when it should have read DEC and JAN. Therefore, this commit changes results in more than just round-off.

The subroutines that were rewritten are:  RDLAI and READLAI.

The only testing that has been done is:
--> 25 vs. 24+1 hour reproducibility
--> writing out XYLAI field from old and new models to confirm they are the same (at least at one time snapshot).
--> running a year to confirm it reads the right months (including a stop/restart during 1st half of a month)