Committing functioning/tested but unpolished code that transforms tracer accumulations into the pre-treated form needed for offline postprocessing.  The major elements of the pre-treating are 1. homogenization so that no accumulations ever need to be divided    by gridbox area during regridding or final scaling, 2. tabulation of denominator indices for ratios, 3. merging of "n" and "s" instances of acc-arrays. Configurations tested: water isotopes, chemistry, Koch aerosols, and the standard 9 tracers.
A new file TRDIAG.f was introduced for grid-independent diagnostics
routines needed for both latlon and cubed-sphere grids.  Routines
like TRACEA et al. will be moved to TRDIAG.f. At this time,
TRDIAG.f is only needed for runs with offline postprocessing.

To ensure that future "specials" added to the existing online print
routines do not need to be duplicated in TRDIAG.f, the online
print routines will be replaced by generic ones free of special cases.
This has to wait until all tracer developers are ok with it.