Adding a non-generic routine REDIST_PACK currently intended only for the parallelization of the atmosphere-ocean coupling routines in OCN_Int_LATLON.f.
The processor assigned to a particular range of latitudes of the ocean
model will, in general, need to receive atmospheric data from other
processors, and vice versa.  Even when the atm/ocean resolutions are the same,
their domain decompositions may differ for load balancing purposes.
The routine works like a "pack" insofar as data from multiple processors
are collected into a global array, but the global array is filled on each
processor, and only where needed.

To collect the contents of a distributed atmospheric array into a
global-sized atmospheric array that is only filled for the atmospheric
J indices overlapping the local ocean domain,

call redist_pack(grid_atm,grid_ocean,atm_array_local,atm_array_global)

To collect in the other direction,

call redist_pack(grid_ocean,grid_atm,ocean_array_local,ocean_array_global)

To-dos:
This initial version is only for 2D arrays.  An F90 interface will be
added. To avoid re-tabulating MPI counts/displacements every time,
they could be passed in a derived type that is computed just once.
The assumption is still made that one grid is just a subdivision
of the other.  The output array need not be global-sized, but that
was the simplest way to start.  Halo updates may be faster than
mpi_alltoallv in some cases.