1. Fixed problem with hanging/crashing model for simulations with dust tracers. Now and then, at a single grid cell, the value for variable sigma exceeds the maximum index value used for the lookup table of dust emission, leading to a violation of the upper boundary of arrays, particularly in simulations with a higher resolution model version. In this fix, sigma, wsgcm, and soilvtrsh are being checked now against their maximum index values for the lookup tables. When one of these variables exceeds the respective maximum value it is set to the maximum value, and a warning message is written to a file. Preferrably, the lookup table for dust emission should be recalculated to allow larger values for sigma. However, since the event doesn't occur often, we don't expect any significant distortion in the statistical moments of the dust emission by keeping the current lookup table for now.
The little utility subroutine check_upper_limit that we put into PBL.f
for above purpose might actually not be at the right place and be
moved to another file, since the routine is independent on any model
physics. We just don't know to what file it belongs best so that there
won't be any circular dependency.

2. Use soilvtrsh instead of wtrsh to calculate emission using the
cubic formula for diagnostic purposes, since the data files for the
legacy emission scheme are only available for the 4x5 version of the
model.