For convenience with cubed-sphere atmosphere runs, ocean C -> atm A regrids are now implemented as ocean C -> ocean A followed by ocean A -> atm A.
For the 1x1.25-degree ocean, results with the 2x2.5 latlon
atmosphere differ only by roundoff, save for slightly
larger differences at the pole related to sin/cos longitudinal
averaging issues.

However, results differ for the 2x2.5 and 4x5 coupled models;
the new implementation avoids a mistake in subroutine
INT_OG2AG_Vector1 (the C -> A regrid was skipped when atm and
ocean resolutions are the same).
Since no template rundecks set UOdrag=1 (use ocean surface
velocity when calculating wind stress), the direct impact
should be limited to ice-covered regions, and the overall
effect is likely to be small in general.