(NOTE TO HYCOM GROUP: your code will no longer compile after the move of dmui,dmvi from the FLUXES module to the ICEDYN_COM module. I can make the changes if you want).
To simplify matters when the ice dynamics grid is not the same
as the atmospheric grid, the ice-ocean momentum flux is passed
directly from the ice to the ocean.  No change to results for
the latlon atmosphere.

Code changes:

Moved dmui,dmvi from the FLUXES module in FLUXES.f to the
ICEDYN_COM module in ICEDYN_DRV.f

In OCN_Interp.f, created routines IG2OG_oceans/OG2IG_oceans which
are the analogs to AG2OG_oceans/OG2AG_oceans called in OCNDYN.f.
OG2IG_oceans will be filled in soon.

Moved the hntrp module from OCN_Int_LATLON.f to OFLUXES.f so as
to be accessible to the cubed-sphere atm configuration.  This module
should eventually go into a modelE "component" directory.

Cleaned up the parts of DYNSI that were forcing ice-to-ocean
fluxes to pass through the atmosphere grid.