Updating the treatment of coastal sea ice advection for the cubed-atmosphere configuration.
Atm cells with <10% ocean coverage do not participate in advection.
This restriction may be removed later if the interactions of ice
mass and ice velocity can be improved for coastal cells.

A better definition of single-cell "inlets" is now made using the ice
dynamics' land-ocean mask interpolated to the advection grid (and the 10%
ocean coverage limit).  Ice is not allowed to enter single-cell inlets
(exits are permitted).