Now that we have template multi-tracer group rundecks, the use of the switch "rad_interact_tr" to determine whether the tracers are used in the final RCOMPX call seems a little dangerous [e.g. if you intended to have the ozone tracer used but the aerosols be climatological]. So, replacing rad_interact_tr entirely by rad_interact_aer (for non chem tracers like aerosols and dust) and rad_interact_chem for gas chem tracers (like O3 and CH4). So, along with that change, I had to change all the "onoff" variables to either onoff_aer or onoff_chem, and the rsign's to rsign_aer's or rsign_chem's.
Also changing this in all template rundecks that had rad_interact_tr
in them. Note that the default in these rundecks remains on (=1), but
in the code itself, default remains off (=0). EXCEPTION: in the dust
rundeck E4TdusF40.R, I kept the default off (rad_interact_aer=0). So
this means it is not consistent (I believe) with how dust is treated
in E4TcadF40.R (the combined tracers deck), but that's how it was in
the first place.

This change was not complicated, but I guess a test of your
configuration's tracer radiative forcing diagnostics may be warranted.
(sorry).