Still in testing are these four new subdaily-daig options:
(1) "oAVG" is L=1 ozone (Ox tracer) averaged over the NSUBDD steps (instead of the existing
    instantaneous). Available for #defined TRACERS_SPECIAL_Shindell

(2) "cAOD" is the equivalent of existing AOD but for clear-sky. These are the DAILY-AVERAGED
    NTRACE species optical depths all written to one file. Available for #defined TRACERS_ON.

(3) "tAOD" is like AOD, but NTRACE species are added up before writing, so "total" optical depth.
    Available for #defined TRACERS_ON.

(4) "ctAOD" the clear-sky version of tAOD from 3 above. Available for #defined TRACERS_ON.