Implemented bundle_interpolation() for the latlon atmosphere configuration, and attempted to unify the cubed- and latlon- atmosphere coding where possible.
Also, added some missing polar averaging to the "bundle" version
of OG2AG_TOC2SST(), and tried to improve readability/clarity
of the bundle versions of all routines.

To within roundoff error, the bundle versions give identical results
after 1 timestep with the 2x2.5 atm + 1x1.25 ocean.  After 1 timestep,
differences start to appear in the clouds routine when the bundled
OG2AG_TOC2SST() is used.  These are assumed to be the result of
instability and relative humidity thresholds, but the number of
differing gridpoints is somewhat larger than previous experience
would suggest.  This is still being investigated.