Calculates lookup table for logarithm of soil dust aerosol emission for a range of index values of the surface wind speed (wsgcm), standard deviation of the surface wind speed (sigma) due to subgrid variability, and surface wind speed threshold of dust emission (soilvtrsh), which depends on soil moisture. In ModelE, the logarithm of dust emission is obtained by interpolating the logarithm of the emission from the lookup table.
The parameters Im, Jm, and Km must be the same as the ones in the
model defined in TRDUST_COM.f. The current parameters in this program
are set to calculate the logarithm of dust emission for a range of
wsgcm, sigma, and soilvtrsh up to around 60 m/s, 60 m/s, and 14 m/s,
respectively. To increase the range increase Im, Jm, and Km in the
module above, re-calculate the lookup table, and adjust the parameters
in the model (Lim, Ljm, Lkm in TRDUST_COM.f), accordingly.