Switching the L=1 Ox, PM2.5, and PM10 sub-daily diags to Surface quantities. But this also affected 7BES and DUTRS options, because:
IN THE PROCESS, I changed the accumulation of general tracer array trcsurf( ),
such that it is only zero-ed out in SURFACE when NS==1, i.e. the first surface
time step. Then, in the accumulation, it it devided by NIsurf. Hence, instead of
representing the quantity from the last surface time step, this is now an average
over the surface time steps. I hope this is physically preferrable?, Gavin put this
in, I believe, but has not had a chance to weigh-in on this. I contacted Christy and
Jan about it, and they defer to him.