Optimizations important when # of cpus == JM-2:
1. A first implementation of "upwind-only" haloing, which
cuts the volume of data exchange by about 40% on average.
A more elegant procedure might push this closer to 50%.
For the time being, this option requires #define UPWIND_HALOS.
Performance testing for various #'s of cpus is still underway.

2. Skipping some duplicate pole-fills and polar-box computations.
If pole-filling is not needed for tracers, or at least can be
skipped for the tracer moments, please let me know and this can
be sped up slightly more.