Added a simple program for array-by-array differencing of two netcdf files.
For now, only the maximum absolute and relative differences and
their locations are reported.  Other statistics can easily be added,
but the ease of visualizing differences when using netcdf format should
limit the need for them.

Generic tools already exist for manipulating netcdf files;
the purpose of this program is only to be customizable.