No longer assuming a latlon grid when setting up lat-lon rectangles for surface sources of 'SF6','CFC11','CFCn','SF6_c'.
The new approach using masks gives the same result, save for
roundoff differences.  As far as I know, this only affects
TRACERS_SPECIAL_Lerner and TRACERS_GASEXCH_ocean_CFC runs.

Todo: store the coordinates of the rectangles and other
info in arrays to avoid repetition of code for the
various source areas.