Changes to be closer to how AR5 runs will go:
(1) removing CH4 surface sources in favor of L=1 overwriting of value from the GHG file of the radiation code.
To achieve this, setting use_rad_ch4=1 in deck.

(2) Replacing CO, NOx, Alkenes, and Paraffin input files with transient versions covering 1850-2000 or
a subset of that period. Note that the model year/day JYEAR/JDAY is what chooses the date for the emissions.
If the model date is outside the range of the input file, the first or last year is used. For this reason, some
transient emissions files start with a field of zeros. For example if a certain source was zero until 1900, it
has an 1890-2000 transient file, where 1890 is zeros, so that it works back to 1850. (As a consequence there
are generally more source files listed for transient case than static 1850 case.)

(3) because of (2) above, changing the model run years from 1949-1951 to 1849-1951. Note that no "radiation years"
like s0_yr or ghg_yr have been modified, however.

(4) NOx aircraft is also transient now, and this special case 3D file has no header, so needs to define rundeck
parameters:
aircraft_Tyr1=1910
aircraft_Tyr2=2000

No code changes for any of this.
If you have problems, let me know (gfaluvegi@giss.nasa.gov)