This is a script for converting modelE fixed format source files to free format.   It is quite a bit more complex than I would like, esp. given that the commercial engine 'SPAG' ought to do all these things.   Essentially the various bits of F90 and CPP interfere with SPAG which expects vanilla F77.
In addition to reformatting the source files, SPAG attempts to use DO / END DO
for loops and replace some GOTO statements with EXIT/CONTINUE statements and
so on.   Also standardizes some of the capitalization.

At the end of the script I pass the free format file through Emacs batch to
take a 1st cut at proper indentation and so on.   This probably will not for
other users, as I have customized Emacs to properly indent continuation
statements mixed inside CPP directives.   Easy enough to turn this bit off -
just comment out the final line in the script.


A simpler script could easily do the pure conversion from fixed -> free, but I
am attempting to generate source files that can be digested by Photran.   At
this point, most files should need very littel hand manipulation to achieve
this goal via this script.   However there are some issues that the script
definitely cannot handle.   The worst of these are unbalanced "if / endif" due
to CPP conditionals.    Because SPAG and Photran treat CPP lines as comments,
constructs like the folowing must be repaired prior to conversion:

#ifdef foo
   if (condition) then
   ...
   elseif (other) then
#else
   if (something) then

   ...
   end if
#endif

To SPAG and Photran, this appears to be missing a closing "end if".   (And it
can be fun in a large file to find out where the actual origination of this
is.)

WARNING:  This file needs to be polished for human consumption, and is not
really intended for general use.